// src/services/star-wars-service/implementations/swapi-service/responses/vehicle-dto.ts
export interface VehicleDto {
    name: string
    model: string
    manufacturer: string
    cost_in_credits: string
    length: string
    max_atmosphering_speed: string
    crew: string
    passengers: string
    cargo_capacity: string
    consumables: string
    vehicle_class: string
    pilots: string[]
    films: string[]
    created: string
    edited: string
    url: string
}